refactor: modularize backend, extract frontend utilities, update script loading, and clean up project structure
