+++ public/index.html
@@ -382,6 +382,8 @@ Option C"
 			<small>SQLite-backed demo. No external deps.</small>
 		</footer>
 
-		<script src="./app.js" defer></script>
+	<script src="../src/frontend/utils/htmlSanitizer.js"></script>
+	<script src="../src/frontend/utils/shuffle.js"></script>
+	<script src="./app.js" defer></script>
 	</body>
 </html>
